+++ src/fileutils.cpp
@@ -6,6 +6,7 @@
 #include <fstream>
 #include <iostream>
 #include <stdlib.h>
+#include <algorithm>
 
 #include "debug.h"
 #include "fileutils.h"
@@ -23,7 +24,9 @@ std::string FileUtils::fileBaseName(const std::string &filename) {
 std::string FileUtils::fileExtension(const std::string &filename) {
     std::string::size_type i = filename.rfind(".");
     if (i != std::string::npos) {
-        return filename.substr(i, filename.length());
+        std::string ext = filename.substr(i, filename.length());
+		std::transform(ext.begin(), ext.end(), ext.begin(), (int(*)(int)) tolower);
+        return ext;
     }
     return "";
 }